requirements.txt
# Core Dependencies
asyncio
aiohttp
pydantic
pydantic-settings

# Network & Packet Processing
scapy
pyshark
dpkt


# Machine Learning & Deep Learning
torch
tensorflow
scikit-learn
xgboost
lightgbm
numpy
pandas

# Graph Neural Networks
torch-geometric
networkx
dgl

# Data Processing
dask
joblib

# Database
sqlalchemy
psycopg2-binary
redis
timescaledb

# API & Web
fastapi
uvicorn[standard]
websockets

# Monitoring & Logging
prometheus-client
structlog
python-json-logger

# Configuration
pyyaml
python-dotenv
configparser

# Security Tools Integration
python-whois
dnspython
cryptography

# Testing
pytest
pytest-asyncio
pytest-cov
pytest-mock

# Utilities
click
rich
tqdm
python-dateutil

# ML Experiment Tracking
mlflow
wandb

# Data Augmentation
imbalanced-learn
smote-variants

# Time Series
statsmodels
prophet

# HTTP Client
httpx
requests

# CLI Enhancements
tabulate
colorama
